Create Minimum Swaps to Sort (#81)
Given an array of n distinct elements. Find the minimum number of swaps required to sort the array in strictly increasing order.
You do not need to read input or print anything. Your task is to complete the function minSwaps() which takes the nums as input parameter and returns an integer denoting the minimum number of swaps required to sort the array. If the array is already sorted, return 0. 

here the solution of above problem.